validator methods refactor
